# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Semantic Growth Index API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfilePlan = typeof UserProfilePlan[keyof typeof UserProfilePlan];


export const UserProfilePlan = {
  free: 'free',
  premium: 'premium',
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  plan: UserProfilePlan;
  sgiScore: number;
  /** @nullable */
  sgiDailyDelta: number | null;
  /** @nullable */
  sgiWeeklyDelta: number | null;
  /** @nullable */
  sgiMonthlyDelta: number | null;
  /** @nullable */
  globalRank: number | null;
  totalUsers: number;
  /** @nullable */
  percentile: number | null;
  /** @nullable */
  rankChange30d: number | null;
  createdAt: string;
}

export interface UserSync {
  clerkId: string;
  email: string;
}

export interface SgiSnapshot {
  id: number;
  score: number;
  conceptualComplexity: number;
  semanticVariety: number;
  interdisciplinaryScore: number;
  reasoningDepth: number;
  originality: number;
  stability: number;
  continuity: number;
  timestamp: string;
}

export interface SemanticMapNode {
  id: string;
  domain: string;
  explorationScore: number;
  messageCount: number;
}

export interface SemanticMapEdge {
  source: string;
  target: string;
  strength: number;
}

export interface SemanticMap {
  nodes: SemanticMapNode[];
  edges: SemanticMapEdge[];
}

export interface DomainStrengths {
  strongAreas: string[];
  developmentAreas: string[];
}

export interface PredictionScenario {
  rank30d: number;
  rank90d: number;
  rank180d: number;
  sgi30d: number;
  sgi90d: number;
  sgi180d: number;
}

export interface Predictions {
  conservative: PredictionScenario;
  realistic: PredictionScenario;
  optimistic: PredictionScenario;
}

export interface LeaderboardEntry {
  rank: number;
  displayName: string;
  sgiScore: number;
  percentile: number;
  /** @nullable */
  rankChange30d: number | null;
  isCurrentUser: boolean;
}

export interface LeaderboardPage {
  entries: LeaderboardEntry[];
  total: number;
  currentUserEntry?: LeaderboardEntry;
}

export interface LeaderboardSummary {
  totalUsers: number;
  averageSgi: number;
  topSgi: number;
  top1PercentThreshold: number;
  top10PercentThreshold: number;
  usersActive7d: number;
}

export interface Badge {
  id: number;
  badgeKey: string;
  name: string;
  description: string;
  earnedAt: string;
}

export type MissionType = typeof MissionType[keyof typeof MissionType];


export const MissionType = {
  weekly: 'weekly',
  monthly: 'monthly',
} as const;

export interface Mission {
  id: number;
  type: MissionType;
  title: string;
  description: string;
  progress: number;
  target: number;
  completed: boolean;
  expiresAt: string;
}

export interface GamificationProfile {
  xp: number;
  level: number;
  streak: number;
  /** @nullable */
  lastActiveDate: string | null;
  badges: Badge[];
  missions: Mission[];
  xpToNextLevel: number;
  levelProgress: number;
}

export interface Recommendation {
  id: number;
  category: string;
  content: string;
  /** @nullable */
  estimatedSgiGain: number | null;
  createdAt: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  /** @nullable */
  sgiDelta: number | null;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  sgiDelta: number | null;
  createdAt: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  /** @nullable */
  sgiDelta: number | null;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface ApiError {
  error: string;
}

export type GetSgiHistoryParams = {
days?: number;
};

export type GetLeaderboardParams = {
limit?: number;
offset?: number;
};